Avoid trying to unlink directory

+++ javascript/bin/lib.js
@@ -6,10 +6,12 @@ function error(msg, exit_code) {
   process.exit(exit_code || 0);
 };
 
-function clear_dir(dir) {
+function clear_dir(dir, ext) {
   var files = fs.readdirSync(dir);
   for (var file of files) {
-    fs.unlinkSync(path.join(dir, file));
+    if (file.slice(-ext.length) === ext) {
+      fs.unlinkSync(path.join(dir, file));
+    }
   };
 };
 
@@ -89,8 +91,8 @@ function _fm_(main = "main", dir = ".", ext = ".fm", parse = fm.lang.parse, show
   if (errors.length === 0 && ext === ".fm") {
     if (!fs.existsSync(".fmc")) fs.mkdirSync(".fmc");
     if (!fs.existsSync(".fml")) fs.mkdirSync(".fml");
-    clear_dir(".fmc");
-    clear_dir(".fml");
+    clear_dir(".fmc", ".fmc");
+    clear_dir(".fml", ".fml");
     for (var name in synt) {
       var code = "";
       code += name + ": ";
